pyproject.toml
[tool.poetry]
name = "bookB116"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[[tool.poetry.source]]
name = 'default'
url = 'https://pypi.tuna.tsinghua.edu.cn/simple'
default = true

[tool.poetry.dependencies]
python = "^3.6.1"

cryptography = "^3.4"
Flask = "^1.1.1"
Flask-Cors = "^3.0.10"
Flask-Login = "^0.5.0"
Flask-SQLAlchemy = "^2.4.1"
Flask-WTF = "^0.14.3"
lxml = "^4.6.2"
mysql-connector-python = "^8.0.23"
PyYAML = "^5.4.1"
python-box = "^5.3.0"
python-dotenv = "^0.15.0"
uWSGI = { version="~2.0.19", optional=true }
Flask-Migrate = "^2.7.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
flake8 = "^3.8.4"
coverage = "^5.5"
autopep8 = "^1.5.5"
taskipy = "^1.6.0"

[tool.poetry.extras]
uwsgi = ["uWSGI"]
ws = ["websocket-client"]

[tool.pytest.ini_options]
testpaths = "tests"
addopts = "-v"
filterwarnings = [
    "ignore::DeprecationWarning"
]

[tool.coverage.run]
branch = true
source = ["bookB116"]
omit = [
    "*email*"
]

[tool.taskipy.tasks]
test-cov = "coverage run -m pytest"
cov-report = "coverage report"

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
